Added traps and battery to maze
