--- Assets/BatteryManager1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 285bd4cf7f7083a42b3795156db51bb2
