+++ Assets/Prefabs/battery-Capsule.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1e064ead77e8b24094e2091261ed01b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
