+++ Assets/Prefabs/traps.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cf109a474488545419c0e493ffdf7eee
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
