+++ Assets/Scripts/BatteryManager1.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f6cf457e7d4460a449e7e2a3f7b3c5dc
