--- Assets/Scripts/BatteryPickup.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f68414a076c49aa4281b76f35ce1b78b
